Refactor overlay structure
