+++ src/components/login/LoginForm.js
@@ -9,7 +9,7 @@ import React, { useContext, useState } from 'react';
 import PropTypes from 'prop-types';
 import { NavigationContext } from 'react-navigation';
 import { ksStyle } from '../../styles/basic/ksBasic';
-import LoginButtonOnPress from '../../modules/login/loginButtonOnPress';
+import LoginWithEmail from '../../modules/login/sendEmailWithAuthCode';
 
 const styles = StyleSheet.create({
   input: {
@@ -38,7 +38,18 @@ function LoginForm({ loginButtonPressed, setError }) {
   const navigation = useContext(NavigationContext);
 
   const [email, setEmail] = useState('justynabasiak@interia.pl');
-  const [pairID, setPairID] = useState('2');
+  const [pairNr, setPairNr] = useState('2');
+
+  async function onPress() {
+    const { status, message } = await LoginWithEmail(email, pairNr);
+    if (status === 200) {
+      loginButtonPressed(false);
+      navigation.navigate('EmailAuth', { email });
+    } else {
+      loginButtonPressed(false);
+      setError({ isError: true, message });
+    }
+  }
 
   return (
     <View>
@@ -56,22 +67,14 @@ function LoginForm({ loginButtonPressed, setError }) {
         style={styles.input}
         placeholder="Nr pary"
         keyboardType="numeric"
-        value={pairID}
-        onChangeText={setPairID}
+        value={pairNr}
+        onChangeText={setPairNr}
         disabled
         placeholderTextColor="rgba(0, 0, 0, 0.2)"
       />
       <TouchableHighlight
         style={styles.login_button_container}
-        onPress={() =>
-          LoginButtonOnPress(
-            email,
-            pairID,
-            loginButtonPressed,
-            setError,
-            navigation
-          )
-        }
+        onPress={() => onPress()}
       >
         <Text style={styles.login_button_text}>Zaloguj</Text>
       </TouchableHighlight>
