+++ src/components/login/authenticationOverlay/index.js
@@ -1,0 +1,3 @@
+import AuthenticationOverlay from './AuthenticationOverlay';
+
+export default AuthenticationOverlay;
