--- src/modules/login/loginButtonOnPress.js
@@ -1,22 +1,0 @@
-import LoginWithEmail from './sendEmailWithAuthCode';
-
-async function LoginButtonOnPress(
-  email,
-  pairNr,
-  loginButtonPressed,
-  setError,
-  navigation
-) {
-  loginButtonPressed(true);
-
-  const { status, message } = await LoginWithEmail(email, pairNr);
-  if (status === 200) {
-    loginButtonPressed(false);
-    navigation.navigate('EmailAuth', { email });
-  } else {
-    loginButtonPressed(false);
-    setError({ isError: true, message });
-  }
-}
-
-export default LoginButtonOnPress;
