+++ src/modules/navigation/AuthStack.js
@@ -1,12 +1,12 @@
 import { createStackNavigator } from 'react-navigation-stack';
 import LoginScreen from '../../views/loginScreen/LoginScreen';
 import { ksStyle } from '../../styles/basic/ksBasic';
-import EmailAuthenticationOverlay from '../../views/loginScreen/EmailAuthenticationOverlay';
+import AuthenticationOverlay from '../../components/login/authenticationOverlay';
 
 const AuthStack = createStackNavigator(
   {
     LoginScreen,
-    EmailAuth: EmailAuthenticationOverlay,
+    EmailAuth: AuthenticationOverlay,
   },
   {
     initialRouteName: 'LoginScreen',
